+++ src/commands/mod.rs
@@ -1,4 +1,7 @@
 // pub mod join;
+pub mod leave;
+pub mod pause;
 pub mod ping;
+pub mod stop;
 // pub mod play;
 // pub mod skip;
