+++ src/commands/ping.rs
@@ -2,6 +2,7 @@ use crate::{check_msg, CommandResult, Context};
 
 #[poise::command(slash_command, guild_only)]
 pub async fn ping(ctx: Context<'_>) -> CommandResult {
-    check_msg(ctx.say("Pong!").await);
+    let reply = poise::CreateReply::default().content("Pong!");
+    check_msg(ctx.send(reply).await);
     Ok(())
 }
